+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Replace leftover install-tools-action with install-binary-action in generated
+  workflow.
+  
 ### Security
 
 - Update actions/setup-go from v1 to v2.1.3 in generated workflows.